feat: replicaset
This is not easily supported because docker compose replicas
do not maintain unique hostnames. I do not remember how we got
it working before, but the flux fake hostname no
longer works, so I have removed it from main. However,
I want to preserve the code in this branch in case
someone wants to take a look or docker compose makes
this easier to support.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>